# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WellnessBridge API For 8 Tables",
        "description": "API documentation for managing health workers",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/children": {
            "get": {
                "tags": [
                    "Children"
                ],
                "summary": "Get paginated list of children",
                "description": "Display a listing of children resources",
                "operationId": "abfd365f803365f1d58541a7a89a31d9",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter records",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Children retrieved successfully."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Children"
                ],
                "summary": "Create a new child record",
                "description": "Store a newly created child resource",
                "operationId": "2dbf2a88620057466bc48e5ff6b81ed0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "gender",
                                    "dob",
                                    "address",
                                    "parentName",
                                    "parentContact"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2018-01-01"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "parentName": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "parentContact": {
                                        "type": "string",
                                        "example": "555-1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Child created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/children/{id}": {
            "get": {
                "tags": [
                    "Children"
                ],
                "summary": "Get child by ID",
                "description": "Display the specified child resource",
                "operationId": "08e159ed3ccd0a7e604febe3d7a09c31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Child ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Child not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Children"
                ],
                "summary": "Update child record",
                "description": "Update the specified child resource",
                "operationId": "0a70fd57ea00075ea89231896a8481ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Child ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2018-01-01"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "parentName": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "parentContact": {
                                        "type": "string",
                                        "example": "555-1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Child updated successfully"
                    },
                    "404": {
                        "description": "Child not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Children"
                ],
                "summary": "Delete child record",
                "description": "Remove the specified child resource",
                "operationId": "117ea39981da1c397240a949cfa7642e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Child ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Child deleted successfully"
                    },
                    "404": {
                        "description": "Child not found"
                    }
                }
            }
        },
...
    },
...
}